# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.ErrorResponse:
    properties:
      error: {}
      message:
        example: Internal Server Error
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  response.SuccessResponse:
    properties:
      data: {}
      message:
        example: data retrieved/created/updated/deleted successfully
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: Dokumentasi API untuk Nutrione
  title: NutriOne API
  version: "1.0"
paths:
...
  /profiles:
    patch:
      consumes:
      - application/json
      description: Access this endpoint to update User profile
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Profile
      tags:
      - Profiles
  /profiles/avatar:
    post:
      consumes:
      - application/json
      description: Access this endpoint to upload user avatar
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload Profile Avatar
      tags:
      - Profiles
...
swagger: "2.0"